pyproject.toml
[project]
name = "skillbot"
version = "0.3.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "MadaraUchiha-314", email = "rohithr31+github@gmail.com" },
]
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.24",
    "click>=8.3.1",
    "fastapi>=0.131.0",
    "langchain-core>=1.2.14",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "python-frontmatter>=1.1.0",
    "uvicorn>=0.41.0",
    "rich>=14.0.0",
    "aiosqlite>=0.22.1",
]
license = { file = "LICENSE" }

[project.scripts]
skillbot = "skillbot.cli.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["skillbot"]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "skillbot.cli.cli"
disallow_untyped_decorators = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.3.0"
version_files = [
    "pyproject.toml:^version",
    "skillbot/__init__.py:^__version__",
]
tag_format = "v$version"
update_changelog_on_bump = true

[dependency-groups]
dev = [
    "commitizen>=4.13.8",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
